chore: renamed auth to tsky
